# backend/src/boardgames_api/models/common.py
from typing import List, Optional

from pydantic import BaseModel, Field


class ProblemDetails(BaseModel):
    """
    Error payload following RFC 9457. Provides structured diagnostic information
    for validation failures, domain errors, and authentication issues.
    """
    type: Optional[str] = Field(
        None, description="A URI reference that identifies the problem type."
    )
    title: str = Field(
        ..., description="A short, human-readable summary of the problem."
    )
    status: int = Field(
        ..., ge=100, le=599, description="HTTP status code generated by the origin server."
    )
    detail: Optional[str] = Field(
        None,
        description="A human-readable explanation specific to this occurrence of the problem.",
    )
    instance: Optional[str] = Field(
        None,
        description="A URI reference that identifies the specific occurrence of the problem.",
    )
    code: Optional[str] = Field(None, description="A machine-readable error code.")
    invalid_params: Optional[List[dict[str, str]]] = Field(
        None,
        description="A list of invalid parameters with details about the validation failure.",
    )


class Paginated(BaseModel):
    """
    Pagination envelope used for listing resources.
    """
    total: int = Field(..., ge=0, description="The total number of items available.")
    limit: int = Field(
        ..., ge=1, le=100, description="The maximum number of items returned per page."
    )
    offset: int = Field(..., ge=0, description="The starting position of the items returned.")
    items: List[dict[str, object]] = Field(
        ..., description="The list of items in the current page."
    )