# openapi.yaml
openapi: 3.1.0

info:
  title: Boardgame Recommender API
  version: "0.2.0"
  description: |
    API for a boardgame recommender used in controlled user studies.
    Provides canonical boardgame metadata and preference-based
    recommendation generation with structured explainability.

servers:
  - url: /api/

tags:
  - name: auth
    description: Session initialization and termination.
  - name: games
    description: Canonical boardgame metadata used by the recommender.
  - name: recommendations
    description: Creation and retrieval of recommendation sessions.

components:

  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id

  schemas:

    ProblemDetails:
      type: object
      description: >
        Error payload following RFC 9457. Provides structured diagnostic information
        for validation failures, domain errors (e.g., unknown game), and authentication issues.
      properties:
        type: { type: string, format: uri }
        title: { type: string, minLength: 1 }
        status: { type: integer, minimum: 100, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri }
        code: { type: string }
        invalid_params:
          type: array
          items:
            type: object
            required: [name, reason]
            properties:
              name: { type: string }
              reason: { type: string }
              code: { type: string }
      unevaluatedProperties: false
      examples:
        - generic:
            summary: Generic RFC 9457 'about:blank' error
            value:
              type: "about:blank"
              title: "Internal Server Error"
              status: 500
              detail: "The server encountered an internal error while processing the request. Please try again later."
        - domain:
            summary: Domain error with specific type + code
            value:
              type: "https://example.com/errors/game-not-found"
              title: "Game not found"
              status: 404
              detail: "No boardgame exists with the specified ID."
              code: "GAME_NOT_FOUND"
        - validation:
            summary: Validation error illustrating structured field errors
            value:
              type: "https://example.com/errors/validation-error"
              title: "Validation failed"
              status: 400
              detail: "One or more input parameters were invalid."
              code: "VALIDATION_ERROR"
              invalid_params:
                - name: "play_context.players"
                  reason: "Value must be ≥ 1."
                  code: "RANGE_MIN"
                - name: "liked_games[0]"
                  reason: "Must match pattern '^[0-9]+$'."
                  code: "FORMAT_INVALID"
        - auth:
            summary: Authentication/authorization error
            value:
              type: "https://example.com/errors/unauthorized"
              title: "Authentication required"
              status: 401
              detail: "A valid session cookie is required to access this resource."
              code: "AUTH_INVALID_SESSION"

    Participant:
      type: object
      description: Representation of a study participant identified by a random ID.
      properties:
        participant_id: { type: string, minLength: 1 }
        study_group: { type: string }
      required: [participant_id, study_group]
      unevaluatedProperties: false

    Paginated:
      type: object
      description: Pagination envelope used for listing resources.
      properties:
        total: { type: integer, minimum: 0 }
        limit: { type: integer, minimum: 1, maximum: 100 }
        offset: { type: integer, minimum: 0 }
        items:
          type: array
          items: { type: object }
      required: [total, limit, offset]
      unevaluatedProperties: false

    BoardGame:
      type: object
      description: |
        Canonical boardgame metadata aligned with the model’s feature set.
      properties:
        id: { type: string, pattern: "^[0-9]+$" }
        title: { type: string, minLength: 1 }
        description: { type: string }
        mechanics:
          type: array
          items: { type: string, minLength: 1 }
          uniqueItems: true
          default: []
        genre:
          type: array
          items: { type: string, minLength: 1 }
          uniqueItems: true
          default: []
        themes:
          type: array
          items: { type: string, minLength: 1 }
          uniqueItems: true
          default: []
        min_players: { type: integer, minimum: 1 }
        max_players: { type: integer, minimum: 1 }
        complexity: { type: number, minimum: 0, maximum: 5 }
        age_recommendation: { type: integer, minimum: 0 }
        num_user_ratings: { type: integer, minimum: 0 }
        avg_user_rating: { type: number, minimum: 0, maximum: 10 }
        year_published: { type: integer, minimum: -3500, maximum: 2021 }
        playing_time_minutes: { type: integer, minimum: 1 }
        image_url: { type: string, format: uri }
        bgg_url: { type: string, format: uri }
      required:
        - id
        - title
        - description
        - mechanics
        - genre
        - themes
        - min_players
        - max_players
        - complexity
        - age_recommendation
        - num_user_ratings
        - year_published
        - playing_time_minutes
        - image_url
        - bgg_url
      unevaluatedProperties: false

    Recommendation:
      type: object
      description: |
        A recommendation entry paired with a structured explanation.
      properties:
        boardgame:
          $ref: "#/components/schemas/BoardGame"

        explanation:
          type: object
          description: Structured explanation of why this game was recommended.
          unevaluatedProperties: false
          required: [type]
          properties:
            type:
              type: string
              enum: ["references", "features"]

            references:
              type: array
              description: Reference-based reasoning using familiar games selected by the backend.
              minItems: 1
              maxItems: 3
              items:
                type: object
                unevaluatedProperties: false
                required: [bgg_id, title]
                properties:
                  bgg_id:
                    type: integer
                    minimum: 1
                  title:
                    type: string
                    minLength: 1

            features:
              type: array
              description: Feature-based reasoning exposing key domain factors used by the model.
              minItems: 1
              maxItems: 3
              items:
                type: object
                unevaluatedProperties: false
                required: [label, category]
                properties:
                  label:
                    type: string
                  category:
                    type: string
                    enum:
                      - mechanic
                      - theme
                      - genre
                      - playtime
                      - complexity
                      - age

              allOf:
                - if:
                    properties:
                      type: { const: "references" }
                  then:
                    required: ["type", "references"]
                - if:
                    properties:
                      type: { const: "features" }
                  then:
                    required: ["type", "features"]
      required: [boardgame, explanation]
      unevaluatedProperties: false

    RecommendationRequest:
      type: object
      description: |
        Preference profile supplied for generating recommendations.
      unevaluatedProperties: false
      properties:
        liked_games:
          type: array
          uniqueItems: true
          default: []
          items: { type: integer, minimum: 1 }

        disliked_games:
          type: array
          uniqueItems: true
          default: []
          items: { type: integer, minimum: 1 }

        play_context:
          type: object
          unevaluatedProperties: false
          properties:
            players: { type: integer, minimum: 1 }
            duration:
              type: string
              enum: ["short", "medium", "long"]

        preference_tags:
          type: array
          uniqueItems: true
          default: []
          items:
            type: string
            enum:
              - strategy
              - cooperative
              - family-friendly
              - low-luck
              - thematic
              - party
              - economic
              - abstract

        avoid_tags:
          type: array
          uniqueItems: true
          default: []
          items:
            type: string
            enum:
              - strategy
              - cooperative
              - family-friendly
              - low-luck
              - thematic
              - party
              - economic
              - abstract

        num_results:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
      required: [num_results]

    RecommendationResponse:
      type: object
      description: |
        A complete recommendation session including input, model version,
        and produced ranked results.
      properties:
        session_id: { type: string, minLength: 1 }
        participant_id: { type: string, minLength: 1 }
        created_at: { type: string, format: date-time }
        intent:
          $ref: "#/components/schemas/RecommendationRequest"
        model_version: { type: string, minLength: 1 }
        experiment_group: { type: string }
        recommendations:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/Recommendation"
      required:
        - session_id
        - participant_id
        - created_at
        - intent
        - model_version
        - recommendations
      unevaluatedProperties: false

paths:

  /auth/session:
    post:
      tags: [auth]
      summary: Create a participant session.
      description: Creates a new session using a study token and issues a session cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              unevaluatedProperties: false
              required: [study_token]
              properties:
                study_token: { type: string, minLength: 1 }
      responses:
        "200":
          description: Session created.
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Participant"
        "400":
          description: Invalid study token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProblemDetails" }

    delete:
      tags: [auth]
      summary: Terminate the current participant session.
      security:
        - sessionCookie: []
      responses:
        "204":
          description: Session deleted.
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProblemDetails" }

  /recommendation:
    post:
      tags: [recommendations]
      summary: Generate recommendations for a participant.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RecommendationRequest" }
      responses:
        "201":
          description: Recommendation session created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RecommendationResponse" }
        "400":
          description: Invalid request.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProblemDetails" }
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProblemDetails" }
        "503":
          description: Recommender unavailable.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProblemDetails" }

  /recommendation/{session_id}:
    get:
      tags: [recommendations]
      summary: Retrieve a stored recommendation session.
      security:
        - sessionCookie: []
      parameters:
        - in: path
          name: session_id
          required: true
          schema: { type: string, minLength: 1 }
      responses:
        "200":
          description: Recommendation session retrieved.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RecommendationResponse" }
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProblemDetails" }
        "404":
          description: Session not found.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProblemDetails" }

  /games:
    get:
      tags: [games]
      summary: List canonical boardgames.
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 10 }

        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }

        - in: query
          name: genre
          style: form
          explode: true
          schema:
            type: array
            items: { type: string, minLength: 1 }

        - in: query
          name: mechanics
          style: form
          explode: true
          schema:
            type: array
            items: { type: string, minLength: 1 }

        - in: query
          name: themes
          style: form
          explode: true
          schema:
            type: array
            items: { type: string, minLength: 1 }

        - in: query
          name: q
          schema: { type: string }
      responses:
        "200":
          description: Boardgames retrieved.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Paginated"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/BoardGame"
        "400":
          description: Invalid pagination parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"

  /games/{bgg_id}:
    get:
      tags: [games]
      summary: Retrieve metadata for a boardgame.
      parameters:
        - in: path
          name: bgg_id
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Boardgame retrieved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardGame"
        "404":
          description: Game not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "400":
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"